Changed text printed in cli/daemon
On line 2 of the Rust file, main.rs, which is written in Rust that is situated within the cli/daemon/src directory, I have modified the str given as an argument to the println! macro from "Hello, world!" to "Hi spez". println! is a macro that prints to the standard output, with a newline.

On all platforms, the newline is the LINE FEED character (\n/U+000A) alone (no additional CARRIAGE RETURN (\r/U+000D)).

This macro uses the same syntax as format!, but writes to the standard output instead. See std::fmt for more information.

The println! macro will lock the standard output on each call. If you call println! within a hot loop, this behavior may be the bottleneck of the loop. To avoid this, lock stdout with io::stdout().lock():.